Create Routes for different pages, Move App to pages/Datacenters
