+++ project/client/src/components/Navbar/Navbar.tsx
@@ -1,12 +1,30 @@
 import React from 'react';
 import { Header, Title, NavList, NavItem } from './Navbar.styles';
+import { useHistory } from "react-router-dom";
+
 
 const Navbar: React.FC = () => {
-  return (
+  const history = useHistory();
+
+  function navHome() {
+    history.push('/');
+  }
+
+  function navDCs() {
+    history.push('/Datacenters');
+  }
+
+  function navJobs() {
+    history.push('/Jobs');
+  }
+
+  return (  
     <Header>
       <Title>My CO2 App</Title>
       <NavList>
-        <NavItem>Home</NavItem>
+        <NavItem onClick={navHome}>Home</NavItem>
+        <NavItem onClick={navDCs}>Datacenters</NavItem>
+        <NavItem onClick={navJobs}>Jobs</NavItem>
         <NavItem>About</NavItem>
         <NavItem>Contact</NavItem>
       </NavList>
