+++ project/client/src/pages/Homepage/Homepage.tsx
@@ -1,0 +1,15 @@
+import React, { useEffect, useState } from 'react';
+import { Content } from './Homepage.styles';
+
+const App: React.FC = () => {
+
+  return (
+    <div>
+      <Content>
+        <p>Put Homepage view here.</p>
+      </Content>
+    </div>
+  );
+}
+
+export default App;
