+++ project/client/src/pages/Jobs/Jobs.styles.tsx
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const Content = styled.div`
+  display: flex;
+  justify-content: space-between;
+  gap: 20px;
+`;
