+++ project/client/src/pages/Jobs/Jobs.tsx
@@ -1,0 +1,15 @@
+import React, { useEffect, useState } from 'react';
+import { Content } from './Jobs.styles';
+
+const App: React.FC = () => {
+  
+  return (
+    <div>
+      <Content>
+        <p>Put jobs view here.</p>
+      </Content>
+    </div>
+  );
+}
+
+export default App;
